requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
fastapi==0.110.0
h11==0.14.0
httptools==0.6.4
idna==3.10
joblib==1.4.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pydantic==2.11.3
pydantic_core==2.33.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
starlette==0.36.3
threadpoolctl==3.6.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.27.0.post1
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
